refactor: ファイル構造をv1ディレクトリに統一
- src/api/auth/login → src/api/v1/auth/login
- test/api/auth/login → test/api/v1/auth/login
- importパスを適切に更新
- APIの統一的なディレクトリ構造を維持
